requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
click==8.1.7
rich==13.7.0
httpx==0.25.2

# LLM and AI
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
jsonschema==4.21.1
numpy==1.26.3
scikit-learn==1.4.0

# Vector Store
weaviate-client==3.25.3

# Observability and Logging
structlog==23.2.0
prometheus-client==0.19.0
python-json-logger==2.0.7

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development
black==23.11.0
ruff==0.1.14
mypy==1.7.1

# Production
gunicorn==21.2.0
redis==5.0.1  # For caching
python-multipart==0.0.6  # For file uploads

# Optional dependencies
# llama-cpp-python==0.2.38  # For GGUF support
# accelerate==0.26.1  # For GPU optimization